Making sure to run CUDA-dependent tests only run when CUDA is available.

+++ test/TorchSharpTest/TestLoadSave.cs
@@ -247,19 +247,21 @@ namespace TorchSharp
         [Fact]
         public void TestSaveLoadGruOnCudaFreezes()
         {
-            // Fails on CUDA
-            if (File.Exists(".model.ts")) File.Delete(".model.ts");
-            var gru = GRU(2, 2, 2);
-            gru.to(DeviceType.CUDA);
-            var params0 = gru.parameters();
-            gru.save(".model.ts");
+            if (torch.cuda.is_available()) {
+                // Fails on CUDA
+                if (File.Exists(".model.ts")) File.Delete(".model.ts");
+                var gru = GRU(2, 2, 2);
+                gru.to(DeviceType.CUDA);
+                var params0 = gru.parameters();
+                gru.save(".model.ts");
 
-            var loadedGru = GRU(2, 2, 2);
-            loadedGru.to(DeviceType.CUDA);
-            loadedGru.load(".model.ts");
-            var params1 = loadedGru.parameters();
-            File.Delete(".model.ts");
-            Assert.Equal(params0, params1);
+                var loadedGru = GRU(2, 2, 2);
+                loadedGru.to(DeviceType.CUDA);
+                loadedGru.load(".model.ts");
+                var params1 = loadedGru.parameters();
+                File.Delete(".model.ts");
+                Assert.Equal(params0, params1);
+            }
         }
     }
 }
